# swagger/swagger.yaml
basePath: /api/v1
consumes:
- application/json
definitions:
...
  User:
    description: User (default view)
    example:
      active: true
      admin: true
      email: john.doe@vulcan.example.com
      firstname: John
      id: 967d9966-b561-4233-bd6f-cac603fd8320
      last_login: "2018-09-07T10:40:52Z"
      lastname: Doe
      observer: true
    properties:
      active:
        description: Active
        example: true
        type: boolean
      admin:
        description: Admin
        example: true
        type: boolean
      email:
        description: Email
        example: john.doe@vulcan.example.com
        type: string
      firstname:
        description: First name
        example: John
        type: string
      id:
        description: User ID
        example: 967d9966-b561-4233-bd6f-cac603fd8320
        type: string
      last_login:
        description: last_login
        example: "2018-09-07T10:40:52Z"
        format: date-time
        type: string
      lastname:
        description: Last name
        example: Doe
        type: string
      observer:
        description: Observer
        example: true
        type: boolean
    title: 'Mediatype identifier: user; view=default'
    type: object
...
host: www.vulcan.example.com
info:
  description: Public API for Vulcan Vulnerability Scan Engine
  title: Vulcan API
  version: ""
paths:
...
  /profile:
    get:
      description: Show profile information for the current authenticated user based
        on the key used to make the request.
      operationId: user#profile
      produces:
      - user
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
      schemes:
      - https
      security:
      - Bearer: []
      summary: profile user
      tags:
      - user
  /stats/mttr:
    get:
      description: Get global MTR statistics.
      operationId: global-stats#mttr
      parameters:
      - description: Maximum date to filter statistics by
        in: query
        name: maxDate
        required: false
        type: string
      - description: Minimum date to filter statistics by
        in: query
        name: minDate
        required: false
        type: string
      produces:
      - mttr
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Mttr'
      schemes:
      - https
      security:
      - Bearer: []
      summary: mttr global-stats
      tags:
      - global-stats
  /teams:
    get:
      description: List all teams in Vulcan.
      operationId: teams#list
      parameters:
      - description: Team tag
        in: query
        name: tag
        required: false
        type: string
      produces:
      - team; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/TeamCollection'
      schemes:
      - https
      security:
      - Bearer: []
      summary: list teams
      tags:
      - teams
    post:
      description: Create a new team.
      operationId: teams#create
      parameters:
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/TeamPayload'
      produces:
      - team
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Team'
      schemes:
      - https
      security:
      - Bearer: []
      summary: create teams
      tags:
      - teams
  /teams/{team_id}:
    delete:
      description: Delete a team.
      operationId: teams#delete
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      schemes:
      - https
      security:
      - Bearer: []
      summary: delete teams
      tags:
      - teams
    get:
      description: Show information about a team.
      operationId: teams#show
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - team
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Team'
      schemes:
      - https
      security:
      - Bearer: []
      summary: show teams
      tags:
      - teams
    patch:
      description: Update information about a team.
      operationId: teams#update
      parameters:
      - description: team ID
        in: path
        name: team_id
        required: true
        type: string
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/TeamUpdatePayload'
      produces:
      - team
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Team'
      schemes:
      - https
      security:
      - Bearer: []
      summary: update teams
      tags:
      - teams
  /teams/{team_id}/assets:
    get:
      description: List all assets from a team.
      operationId: assets#list
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - listassetentry; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ListassetentryCollection'
      schemes:
      - https
      security:
      - Bearer: []
      summary: list assets
      tags:
      - assets
    post:
      description: "Creates assets in bulk mode.\n\t\t\tThis operation accepts an
        array of assets and an optional array of group identifiers, and returns an
        array of successfully created assets.\n\t\t\tIf no groups are specified, assets
        will be added to the team's Default group.\n\t\t\tIf one of the specified
        assets already exists for the team but is currently not associated with the
        requested groups, the association is created.\n\t\t\tIf for any reason, the
        creation of an asset fails, an error message will be returned referencing
        the failed asset and the entire operation will be rolled back.\n\t\t\t---\n\t\t\tValid
        asset types:\n\t\t\t- AWSAccount\n\t\t\t- DomainName\n\t\t\t- Hostname\n\t\t\t-
        IP\n\t\t\t- IPRange\n\t\t\t- DockerImage\n\t\t\t- WebAddress\n\t\t\t- GitRepository\n\t\t\t---\n\t\t\tIf
        the asset type is informed, then Vulcan will use that value to create the
        new asset.\n\t\t\tOtherwise, Vulcan will try to automatically discover the
        asset type.\n\t\t\tNotice that this may result in Vulcan creating more than
        one asset.\n\t\t\tFor instance, an user trying to create an asset for \"vulcan.example.com\",
        without specifying the asset type, will end up with two assets created:\n\t\t\t-
        vulcan.example.com (DomainName) and\n\t\t\t- vulcan.example.com (Hostname)."
      operationId: assets#create
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CreateAssetPayload'
      produces:
      - asset; type=collection
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/AssetCollection'
      schemes:
      - https
      security:
      - Bearer: []
      summary: create assets
      tags:
      - assets
  /teams/{team_id}/assets/{asset_id}:
    delete:
      description: Delete an asset.
      operationId: assets#delete
      parameters:
      - description: Asset ID
        in: path
        name: asset_id
        required: true
        type: string
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      schemes:
      - https
      security:
      - Bearer: []
      summary: delete assets
      tags:
      - assets
    get:
      description: Describe an asset.
      operationId: assets#show
      parameters:
      - description: Asset ID
        in: path
        name: asset_id
        required: true
        type: string
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - asset
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Asset'
      schemes:
      - https
      security:
      - Bearer: []
      summary: show assets
      tags:
      - assets
    patch:
      description: Update an asset.
      operationId: assets#update
      parameters:
      - description: Asset ID
        in: path
        name: asset_id
        required: true
        type: string
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/AssetUpdatePayload'
      produces:
      - asset
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Asset'
      schemes:
      - https
      security:
      - Bearer: []
      summary: update assets
      tags:
      - assets
  /teams/{team_id}/assets/multistatus:
    post:
      description: "Creates assets in bulk mode (MultiStatus).\n\t\t\tThis operation
        is similar to the \"Create Assets in Bulk Mode\", with 2 main differences:\n\t\t\t-
        This endpoint is not atomic. Each asset creation request will succeed or fail
        indenpendently of the other requests.\n\t\t\t- This endpoint will return an
        array of AssetResponse in the following way:\n\t\t\t\t· For each asset with
        specified type, returns an AssetResponse indicating the success or failure
        for its creation.\n\t\t\t\t· For each asset with no type specified and successfully
        created, returns one AssetResponse for each auto detected asset.\n\t\t\t\t·
        For each asset with no type specified which its creation produced an error,
        returns one AssetResponse indicating the failure for the creation of its detected
        assets without specifying which exact type failed.\n\t\t\tIn the case of all
        assets being successfully created, this endpoint will return status code 201-Created.
        \n\t\t\tOtherwise, it will return a 207-MultiStatus code, indicating that
        at least one of the requested operations failed.\t\n\t\t"
      operationId: assets#createMultiStatus
      parameters:
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/CreateAssetPayload'
      produces:
      - assetresponse; type=collection
      responses:
        "201":
          description: 'Created: All assets were created with success.'
          schema:
            $ref: '#/definitions/AssetresponseCollection'
        "207":
          description: 'Multiple Status: At least one of the assets failed to be created.'
          schema:
            $ref: '#/definitions/AssetresponseCollection'
      schemes:
      - https
      security:
      - Bearer: []
      summary: createMultiStatus assets
      tags:
      - assets
  /teams/{team_id}/findings:
    get:
      description: List all findings from a team.
      operationId: findings#list findings
      parameters:
      - description: Allows to get findings list at a specific date (incompatible
          and preferential to min and max date params)
        in: query
        name: atDate
        required: false
        type: string
      - description: Allows to get findings list for a specific asset identifier
        in: query
        name: identifier
        required: false
        type: string
      - description: Allows to get findings list until a specific date
        in: query
        name: maxDate
        required: false
        type: string
      - description: Findings maximum score
        in: query
        name: maxScore
        required: false
        type: number
      - description: Allows to get findings list from a specific date
        in: query
        name: minDate
        required: false
        type: string
      - description: Findings minimum score
        in: query
        name: minScore
        required: false
        type: number
      - description: Requested page
        in: query
        name: page
        required: false
        type: number
      - description: Requested page size
        in: query
        name: size
        required: false
        type: number
      - description: 'Sorting criteria. Supported fields: score, -score (for descending
          order)'
        in: query
        name: sortBy
        required: false
        type: string
      - description: Findings Status
        in: query
        name: status
        required: false
        type: string
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - findings_list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Findings_list'
      schemes:
      - https
      security:
      - Bearer: []
      summary: list findings findings
      tags:
      - findings
  /teams/{team_id}/findings/{finding_id}:
    get:
      description: Find a finding.
      operationId: findings#Find finding
      parameters:
      - description: Finding ID
        in: path
        name: finding_id
        required: true
        type: string
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - finding
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Finding'
      schemes:
      - https
      security:
      - Bearer: []
      summary: Find finding findings
      tags:
      - findings
  /teams/{team_id}/findings/{finding_id}/overwrite:
    get:
      description: List Finding Overwrites.
      operationId: findings#List Finding Overwrites
      parameters:
      - in: path
        name: finding_id
        required: true
        type: string
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - finding_overwrite; type=collection
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Finding_overwriteCollection'
      schemes:
      - https
      security:
      - Bearer: []
      summary: List Finding Overwrites findings
      tags:
      - findings
    post:
      description: Overwrite data for a specific finding.
      operationId: findings#Submit a Finding Overwrite
      parameters:
      - in: path
        name: finding_id
        required: true
        type: string
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      - in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/FindingOverwritePayload'
      produces:
      - findings_list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Findings_list'
      schemes:
      - https
      security:
      - Bearer: []
      summary: Submit a Finding Overwrite findings
      tags:
      - findings
  /teams/{team_id}/findings/issues:
    get:
      description: List number of findings and max score per issue.
      operationId: findings#list findings issues
      parameters:
      - description: Allows to get issues list at a specific date
        in: query
        name: atDate
        required: false
        type: string
      - description: Allows to get findings list until a specific date
        in: query
        name: maxDate
        required: false
        type: string
      - description: Allows to get findings list from a specific date
        in: query
        name: minDate
        required: false
        type: string
      - description: Requested page
        in: query
        name: page
        required: false
        type: number
      - description: Requested page size
        in: query
        name: size
        required: false
        type: number
      - description: 'Sorting criteria. Supported fields: score, findings_count (use
          - for descending order. E.g.: -score)'
        in: query
        name: sortBy
        required: false
        type: string
      - description: Findings Status
        in: query
        name: status
        required: false
        type: string
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - findings_issues_list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Findings_issues_list'
      schemes:
      - https
      security:
      - Bearer: []
      summary: list findings issues findings
      tags:
      - findings
  /teams/{team_id}/findings/issues/{issue_id}:
    get:
      description: Find all findings from a team and issue.
      operationId: findings#Find findings from a Issue
      parameters:
      - description: Allows to get findings list at a specific date (incompatible
          and preferential to min and max date params)
        in: query
        name: atDate
        required: false
        type: string
      - description: Issue ID
        in: path
        name: issue_id
        required: true
        type: string
      - description: Allows to get findings list until a specific date
        in: query
        name: maxDate
        required: false
        type: string
      - description: Findings maximum score
        in: query
        name: maxScore
        required: false
        type: number
      - description: Allows to get findings list from a specific date
        in: query
        name: minDate
        required: false
        type: string
      - description: Findings minimum score
        in: query
        name: minScore
        required: false
        type: number
      - description: Requested page
        in: query
        name: page
        required: false
        type: number
      - description: Requested page size
        in: query
        name: size
        required: false
        type: number
      - description: 'Sorting criteria. Supported fields: score, -score (for descending
          order)'
        in: query
        name: sortBy
        required: false
        type: string
      - description: Findings Status
        in: query
        name: status
        required: false
        type: string
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - findings_list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Findings_list'
      schemes:
      - https
      security:
      - Bearer: []
      summary: Find findings from a Issue findings
      tags:
      - findings
  /teams/{team_id}/findings/targets:
    get:
      description: List number of findings and max score per target.
      operationId: findings#list findings targets
      parameters:
      - description: Allows to get targets list at a specific date
        in: query
        name: atDate
        required: false
        type: string
      - description: Allows to get findings list until a specific date
        in: query
        name: maxDate
        required: false
        type: string
      - description: Allows to get findings list from a specific date
        in: query
        name: minDate
        required: false
        type: string
      - description: Requested page
        in: query
        name: page
        required: false
        type: number
      - description: Requested page size
        in: query
        name: size
        required: false
        type: number
      - description: 'Sorting criteria. Supported fields: score, findings_count (use
          - for descending order. E.g.: -score)'
        in: query
        name: sortBy
        required: false
        type: string
      - description: Findings Status
        in: query
        name: status
        required: false
        type: string
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      produces:
      - findings_targets_list
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Findings_targets_list'
      schemes:
      - https
      security:
      - Bearer: []
      summary: list findings targets findings
      tags:
      - findings
...
  /teams/{team_id}/groups/{group_id}/assets/{asset_id}:
    delete:
      description: Remove an asset from a group.
      operationId: asset-group#delete
      parameters:
      - description: Asset ID
        in: path
        name: asset_id
        required: true
        type: string
      - description: Group ID
        in: path
        name: group_id
        required: true
        type: string
      - description: Team ID
        in: path
        name: team_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      schemes:
      - https
      security:
      - Bearer: []
      summary: delete asset-group
      tags:
      - asset-group
...
produces:
- application/json
- application/xml
- application/gob
- application/x-gob
...
schemes:
- https
securityDefinitions:
  Bearer:
    in: header
    name: authorization
    type: apiKey
swagger: "2.0"